return: the value of a bit at a given index

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -11,7 +11,7 @@ int get_bit(unsigned long int n, unsigned int index)
 {
 	double rem = n * 0.5;
 	long int num = n * 0.5;
-	unsigned int i;
+	long unsigned int i;
 	char s[32];
 
 		for (i = 0; rem; i++)
